gen_makefile: indent whole file with spaces
Use spaces as indent for the Makefile snippets and replace them with
tabs just before writing the Makefile. This way we can indent the whole
file with spaces. Instead of mixing spaces and tabs, and frequently
having spaces where tabs would need to be by accident, which then
results in a broken Makefile.

Use two spaces for the tabs, as in the vim modeline at the bottom.

Change-Id: I42eb2dda0f1f48d4a8ea2e1e59f9cfd8df7af800