gen_makefile: support symlinks for src + src_proj
Support the src and src_proj directories to be symlinks by checking if
these paths exist only (-e), instead of checking if these are
directories (-d).

Add a second check to detect broken symlinks and inform the user about
those, so it does not fail later on with a more confusing error. This
may happen if the user runs testenv inside podman, and has a symlink
pointing to a directory that is not mounted inside podman.

Related: OS#6494
Change-Id: Ic5ffa132b29e0a2ca78b02e4c8c05492471b3966