gen_makefile: add --autoreconf-in-src-copy arg
When this new argument is set, do not run "autoreconf -fi" directly in
the source directory. Instead, make a copy of all relevant sources in
the git repository (everything except .git and files that are matched
by gitignore) right before working with the source dir. Then use that
copy instead.

This avoids the problem that ./configure refuses to run if it has
already been executed in the source dir. (It aborts and asks the user to
run "make distclean" first, which is especially annoying if it has to be
done in multiple source dirs before being able to build successfully.)

Copying the sources is very fast due to generating a list of files that
need to be copied first, and due to using rsync which skips the copy if
the file size and timestamp already match. On my machine the copying is
not noticable except for the additional prints:

  updating src_copy: /home/user/.cache/osmo-ttcn3-testenv/host/make2/src_copy/osmo-mgw/
  updating version: 1.13.1.38-78ee9

Change-Id: I18ac50e3441df81e1fe7d8d5321df7e80ab9c650