gen_makefile: {make,src}_dir: use absolute paths
Fix generating wrong relative paths when the make dir (-m) is in the
same directory as gen_makefile.py.

This was too easy to get wrong with relative paths as shown in the
previous patches. I've added tests to mitigate this and switched to
absolute paths to make the logic easier to follow.

Fix for:

  $ ./gen_makefile.py -m build
  $ make -C build libosmocore
  …
  cd build/../src/libosmocore; autoreconf -fi
  /bin/sh: 1: cd: can't cd to build/../src/libosmocore

Fixes: 8ea5c0ce ("gen_makefile: fix building open5gs")
Change-Id: I8bb2af7ffb902fd690e83e36739c2b4415b9ab80