gen_makefile: remove --auto-distclean parameter
The purpose of this parameter was to run "make distclean" automatically
if a source directory was already configured for an in-tree build.
osmo-dev always does out-of-tree builds and so ./configure complains:

  configure: error: source directory already configured; run "make distclean" there first

--auto-distclean worked most of the time, but not always.

The --autoreconf-in-src-copy (-A) parameter makes this obsolete as it
solves the same problem but with a better method. When -A is set,
autoreconf and configure run in a copy of the source directory which
avoids all conflicts with an in-tree build in the original source
directory without the need to remove artifacts from the in-tree build.
This has been in use by testenv in osmo-ttcn3-hacks for a while and is
working reliably.

Change-Id: I28378e6d2cafd209adcfae6b85a032d73ed14cce