gen_makefile: support building PyHSS
Support cloning and building PyHSS and "interesting" dependencies (that
we might want to patch while hacking on PyHSS). Other dependencies such
as sqlalchemy get installed via pip into a venv that is currently shared
by all python projects that osmo-dev can build. We can change this later
on to use multiple venvs or change the dependencies that get built from
source, if needed.

Python projects get built with "python3 -m build" into a whl file, and
then installed into the venv with "pip install".

The above works with projects that have a pyproject.toml. PyHSS
currently doesn't have this yet, but this patchset adds one:
https://github.com/nickvsnetworking/pyhss/pull/258

python-venv-requirements.txt in this patch is a combination of the
relevant dependencies from PyHSS and dependencies that get built from
source.

Related: OS#6862
Change-Id: If40c9e8ea07c9f6c7d379f6d5ff659e95165e4ae