gen_makefile: erlang: fix missing clone and deps
Add a dummy configure target for erlang that ensures the git clone and
building of dependencies is done before attempting to build erlang
projects. A dummy target is used to make the logic consistent with
autotools and meson.

This is in preparation for building python libraries for pyhss, which
will work similarly.

Change-Id: I627e78f9317d99755cfad5d5c8526c3b7f1b3a27