gen_makefile: add --targets
Add an option to specify the targets to include in the Makefile, instead
of always including all of them. This option will be used by testenv to
check early if there are any unknown targets, and to make running the
generated Makefile faster.

Running the Makefile is currently slowed down when having a lot of
projects cloned in osmo-dev's src dir, because regardless of the target
to be built, make is looking for all source files for all projects while
parsing the Makefile. For each project we have:

  libosmocore_configure_files := $(shell find -L …/src/libosmocore …)
  libosmocore_files := $(shell find -L …/src/libosmocore …)

Changing ":=" to "=" doesn't make it faster either, make still executes
the find commands while parsing target lines such as:

  .make.libosmocore.build: .make.libosmocore.configure $(libosmocore_files)

Related: osmo-ttcn3-hacks Ib2d8802b305f145d27aab3c1cc3129709b93d93d
Change-Id: I89bb60e94dd03606dbba5a3609d5e1a95993af5b