gen_makefile: turn projects_deps into a dict
Let read_projects_deps() return a dictionary instead of a list of
tuples. This allows looking up entries in the dictionary by project
name, which I'm making use of in a future patch.

While using a dictionary here makes more sense, I assume this was
implemented as list of tuples because with earlier (long EOL) python3
versions, the order of entries in the dictionary were not stable. From
https://docs.python.org/3/library/stdtypes.html#dict:
> Changed in version 3.7: Dictionary order is guaranteed to be insertion
> order. This behavior was an implementation detail of CPython from 3.6.

Change-Id: I74c53f4d6dda791c5d9e14f2274260f0e8cbbad2